Better flush, prune and purge commands

+++ src/commands/Moderation/purge.js
@@ -1,18 +1,23 @@
 exports.run = async (bot, msg, args) => {
   const count = parseInt(args[0]) || 1
 
-  try {
-    msg.delete()
-    const messages = await msg.channel.fetchMessages({
-      limit: Math.min(count, 100),
-      before: msg.id
-    })
-    await Promise.all(messages.map(m => m.delete()))
-    const m = await msg.channel.send(`${SUCCESS}Purged \`${messages.size}\` messages.`)
-    return m.delete(2000)
-  } catch (err) {
-    console.error(err)
+  let messages = await msg.channel.fetchMessages({
+    limit: Math.min(count, 100),
+    before: msg.id
+  })
+
+  if (!msg.guild || !msg.channel.permissionsFor(msg.guild.me).has('MANAGE_MESSAGES')) {
+    messages = messages.filter(m => m.author.id === bot.user.id)
   }
+
+  if (!messages.size) {
+    return msg.error('There are no messages that can be deleted by user!')
+  }
+
+  await msg.edit(`${PROGRESS}Purging ${messages.size} message(s)\u2026`)
+  await Promise.all(messages.map(m => m.delete()))
+
+  return msg.success(`Purged \`${messages.size}\` messages.`)
 }
 exports.info = {
   name: 'purge',
